For symmetry, removed the fixed default incrementOffset from TimestampIncrementingOffset

+++ src/test/java/io/aiven/connect/jdbc/source/TimestampIncrementingCriteriaTest.java
@@ -77,7 +77,7 @@ public class TimestampIncrementingCriteriaTest {
             criteria = this.criteria;
         }
         final TimestampIncrementingOffset offset = criteria.extractValues(schema, record, null);
-        assertEquals(expected, offset.getIncrementingOffset());
+        assertEquals(expected, offset.getIncrementingOffset().longValue());
     }
 
     @Test
